refactor(frontend): create new components
