+++ src/app/manifest.ts
