+++ src/app/register/page.tsx
@@ -1,0 +1,11 @@
+import NavBar from '@/components/NavBar'
+import RegisterForm from '@/components/auth/RegisterForm'
+
+export default function RegisterPage() {
+  return (
+    <>
+      <NavBar title="Register" />
+      <RegisterForm />
+    </>
+  )
+}
